+++ src/Cirreum.Runtime.Wasm.Msal/Authentication/Enrichment/ExternalGraphUserProfileEnricher.cs
@@ -29,7 +29,6 @@ public sealed class ExternalGraphUserProfileEnricher(
 		"displayName",
 		"givenName",
 		"surname",
-		"mailNickname",
 		"mail",
 		"mobilePhone",
 		"preferredLanguage",