+++ src/Cirreum.Runtime.Wasm.Msal/Authentication/Enrichment/GraphExtendedUserProfileEnricher.cs
@@ -32,7 +32,6 @@ public class GraphExtendedUserProfileEnricher(
 		"displayName",
 		"givenName",
 		"surname",
-		"mailNickname",
 		"birthday",
 		"mail",
 		"mobilePhone",